+++ src/entrypoints/injectedButton.content/components/ImportCsvForm.tsx
@@ -7,14 +7,15 @@ import { useForm } from 'react-hook-form';
 import * as yup from 'yup';
 
 import ColumnSelect from './ColumnSelect';
-import { getCsvColumns } from '../utils/csv';
-import { parseCsv } from '../utils/parse';
-import type { ParsedRow } from '../utils/parse';
-import useAsyncFn from '../utils/useAsyncFn';
+import { getCsvColumns } from '../../../utils/csv';
+import useAsyncFn from '../../../utils/useAsyncFn';
+import { parseCsv } from '../parse';
+import type { ParsedRow } from '../parse';
 
 export type ImportCsvFormValues = {
   files: FileList,
   nameColumn: string,
+  setColumn?: string,
   quantityColumn?: string,
   foilColumn?: string,
   priceColumn?: string,
@@ -25,6 +26,7 @@ const validationSchema: yup.ObjectSchema<ImportCsvFormValues> = yup.object({
     .required(i18n.t('injectedButton.modal.importCsvForm.files.required')),
   nameColumn: yup.string()
     .required(i18n.t('injectedButton.modal.importCsvForm.nameColumn.required')),
+  setColumn: yup.string(),
   quantityColumn: yup.string(),
   foilColumn: yup.string(),
   priceColumn: yup.string(),
@@ -36,7 +38,15 @@ type ImportCsvFormProps = {
 
 function ImportCsvForm({ onSubmit }: ImportCsvFormProps) {
   const [{ value: columns, loading, error }, getColumns] = useAsyncFn(getCsvColumns);
-  const { control, register, handleSubmit, watch, setError } = useForm<ImportCsvFormValues>({
+  const {
+    control,
+    register,
+    handleSubmit,
+    watch,
+    setError,
+    clearErrors,
+    formState: { errors: formErrors, touchedFields },
+  } = useForm<ImportCsvFormValues>({
     resolver: yupResolver(validationSchema),
   });
 
@@ -44,6 +54,7 @@ function ImportCsvForm({ onSubmit }: ImportCsvFormProps) {
     try {
       const res = await parseCsv(data.files[0], {
         name: data.nameColumn,
+        set: data.setColumn,
         quantity: data.quantityColumn,
         isFoil: data.foilColumn,
         price: data.priceColumn,
@@ -51,8 +62,8 @@ function ImportCsvForm({ onSubmit }: ImportCsvFormProps) {
       onSubmit(res);
     }
     catch (e) {
-      if (!(e instanceof Error)) setError('root', { message: 'Unknown error.' });
-      else setError('root', e);
+      console.error('[cardmarket-bulk-import] Failed to Submit CSV file.', e);
+      setError('root', { message: i18n.t('injectedButton.modal.importCsvForm.error') });
     }
   };
 
@@ -63,12 +74,16 @@ function ImportCsvForm({ onSubmit }: ImportCsvFormProps) {
   }, [files, getColumns]);
 
   useEffect(() => {
-    if (error) console.error('[cardmarket-bulk-import] Failed to parse Csv file.', error);
-  }, [error]);
+    if (!error) {
+      clearErrors('files');
+      return;
+    };
+    setError('files', { message: i18n.t('injectedButton.modal.importCsvForm.files.invalid') });
+    console.error('[cardmarket-bulk-import] Failed to parse CSV file.', error);
+  }, [error, clearErrors, setError]);
 
   let columnsEl = null;
   if (loading) columnsEl = (<Spinner />);
-  else if (error) columnsEl = (<h5>Invalid file</h5>);
   else if (columns) columnsEl = (
     <>
       <ColumnSelect<ImportCsvFormValues>
@@ -76,6 +91,13 @@ function ImportCsvForm({ onSubmit }: ImportCsvFormProps) {
         formId="importCsvForm.nameColumn"
         name="nameColumn"
         label={i18n.t('injectedButton.modal.importCsvForm.nameColumn.label')}
+        options={columns}
+      />
+      <ColumnSelect<ImportCsvFormValues>
+        control={control}
+        formId="importCsvForm.setColumn"
+        name="setColumn"
+        label={i18n.t('injectedButton.modal.importCsvForm.setColumn.label')}
         options={columns}
       />
       <ColumnSelect<ImportCsvFormValues>
@@ -108,15 +130,30 @@ function ImportCsvForm({ onSubmit }: ImportCsvFormProps) {
   return (
     <Form noValidate onSubmit={handleSubmit(submitFn)}>
       <Stack gap={2}>
-        <Form.Group controlId="importCsvForm.Files">
+        <Form.Group controlId="importCsvForm.files">
           <Form.Label>{ i18n.t('injectedButton.modal.importCsvForm.files.label') }</Form.Label>
           <Form.Control
             type="file"
             accept=".csv"
             {...register('files')}
+            isInvalid={touchedFields.files && !!formErrors.files}
           />
+          <Form.Control.Feedback type="invalid">
+            { formErrors.files?.message }
+          </Form.Control.Feedback>
         </Form.Group>
         { columnsEl }
+        {
+          formErrors.root && (
+            <>
+              { /* Bootstrap invalid-feedback requires an .is-invalid sibling */ }
+              <div className="is-invalid d-none" />
+              <div className="invalid-feedback mt-0">
+                { formErrors.root.message }
+              </div>
+            </>
+          )
+        }
       </Stack>
     </Form>
   );